[Kerberos] Move classes from test support package.
